chore: move deployEntity logic from ServiceImpl to a deployer port. Removed Service.ts and types associated are moved to a unique source deployment-types.ts
