+++ content/src/service/synchronization/deployRemoteEntity.ts
@@ -7,8 +7,8 @@
 import { streamToBuffer } from '@dcl/catalyst-storage/dist/content-item'
 import { AuthChain } from '@dcl/schemas'
 import { downloadEntityAndContentFiles } from '@dcl/snapshots-fetcher'
+import { DeploymentContext, isInvalidDeployment, LocalDeploymentAuditInfo } from '../../deployment-types'
 import { AppComponents } from '../../types'
-import { DeploymentContext, isInvalidDeployment, LocalDeploymentAuditInfo } from '../Service'
 
 const requestMaxRetries = 10
 const requestRetryWaitTime = 1000
