+++ content/test/integration/E2EAssertions.ts
@@ -5,7 +5,7 @@ import assert from 'assert'
 import { DeploymentData } from 'dcl-catalyst-client'
 import { EntityContentItemReference } from 'dcl-catalyst-commons'
 import { Response } from 'node-fetch'
-import { Deployment } from '../../src/logic/deployment-types'
+import { Deployment } from '../../src/deployment-types'
 import { getDeployments } from '../../src/logic/deployments'
 import { FailedDeployment, FailureReason } from '../../src/ports/failedDeployments'
 import { EntityVersion } from '../../src/types'
