+++ content/test/integration/TestProgram.ts
@@ -2,13 +2,12 @@ import { Entity, EntityType } from '@dcl/schemas'
 import { ILoggerComponent, Lifecycle } from '@well-known-components/interfaces'
 import { ContentClient, DeploymentData } from 'dcl-catalyst-client'
 import fetch from 'node-fetch'
+import { AuditInfo, Deployment, DeploymentOptions, isInvalidDeployment } from '../../src/deployment-types'
 import { EnvironmentConfig } from '../../src/Environment'
-import { AuditInfo, Deployment, DeploymentOptions } from '../../src/logic/deployment-types'
 import { getDeployments } from "../../src/logic/deployments"
 import * as synchronization from '../../src/logic/synchronization'
 import { FailedDeployment } from '../../src/ports/failedDeployments'
 import { main } from '../../src/service'
-import { isInvalidDeployment } from '../../src/service/Service'
 import { AppComponents } from '../../src/types'
 import { deleteFolderRecursive } from './E2ETestUtils'
