+++ content/test/integration/controller/deployment-pagination.spec.ts
@@ -1,8 +1,8 @@
 import { EntityType } from '@dcl/schemas'
 import { fetchJson } from 'dcl-catalyst-commons'
 import { DeploymentField } from '../../../src/controller/Controller'
+import { DeploymentOptions, SortingField, SortingOrder } from '../../../src/deployment-types'
 import { EnvironmentConfig } from '../../../src/Environment'
-import { DeploymentOptions, SortingField, SortingOrder } from '../../../src/logic/deployment-types'
 import { toQueryParams } from '../../../src/logic/toQueryParams'
 import { PointerChangesFilters } from '../../../src/service/pointers/types'
 import { makeNoopValidator } from '../../helpers/service/validations/NoOpValidator'
