+++ content/test/integration/logic/snapshots.spec.ts
@@ -1,9 +1,9 @@
 import { EntityType } from '@dcl/schemas'
 import { IBaseComponent } from '@well-known-components/interfaces'
+import { DeploymentContext } from '../../../src/deployment-types'
 import * as snapshotQueries from '../../../src/logic/database-queries/snapshots-queries'
 import { generateSnapshotsInMultipleTimeRanges } from '../../../src/logic/snapshots'
 import * as timeRangeLogic from '../../../src/logic/time-range'
-import { DeploymentContext } from '../../../src/service/Service'
 import { AppComponents } from '../../../src/types'
 import { makeNoopServerValidator, makeNoopValidator } from '../../helpers/service/validations/NoOpValidator'
 import { setupTestEnvironment, testCaseWithComponents } from '../E2ETestEnvironment'
