+++ content/test/integration/ports/deployer/deployment-filters.spec.ts
@@ -1,12 +1,10 @@
 import { EntityType } from '@dcl/schemas'
-import { AuditInfo, DeploymentFilters } from '../../../../src/logic/deployment-types'
-import { getDeployments } from '../../../../src/logic/deployments'
 import {
-  DeploymentContext,
-  DeploymentResult,
+  AuditInfo, DeploymentContext, DeploymentFilters, DeploymentResult,
   isInvalidDeployment,
   isSuccessfulDeployment
-} from '../../../../src/service/Service'
+} from '../../../../src/deployment-types'
+import { getDeployments } from '../../../../src/logic/deployments'
 import { AppComponents } from '../../../../src/types'
 import { makeNoopServerValidator, makeNoopValidator } from '../../../helpers/service/validations/NoOpValidator'
 import { setupTestEnvironment, testCaseWithComponents } from '../../E2ETestEnvironment'
