+++ content/test/integration/service/snapshots/SnapshotManager.spec.ts
@@ -5,9 +5,9 @@ import { createConfigComponent } from '@well-known-components/env-config-provide
 import { ILoggerComponent } from '@well-known-components/interfaces'
 import { createLogComponent } from '@well-known-components/logger'
 import { inspect } from 'util'
+import { isSuccessfulDeployment } from '../../../../src/deployment-types'
 import { EnvironmentBuilder } from '../../../../src/Environment'
 import { stopAllComponents } from '../../../../src/logic/components-lifecycle'
-import { isSuccessfulDeployment } from '../../../../src/service/Service'
 import { SnapshotMetadata } from '../../../../src/service/snapshots/SnapshotManager'
 import { AppComponents } from '../../../../src/types'
 import { makeNoopServerValidator, makeNoopValidator } from '../../../helpers/service/validations/NoOpValidator'
