+++ content/test/unit/ports/activeEntities.spec.ts
@@ -6,8 +6,8 @@ import { createLogComponent } from '@well-known-components/logger'
 import { createTestMetricsComponent } from '@well-known-components/metrics'
 import { HTTPProvider } from 'eth-connect'
 import ms from 'ms'
+import { Deployment } from '../../../src/deployment-types'
 import { DEFAULT_ENTITIES_CACHE_SIZE, Environment, EnvironmentConfig } from '../../../src/Environment'
-import { Deployment } from '../../../src/logic/deployment-types'
 import * as deployments from '../../../src/logic/deployments'
 import { metricsDeclaration } from '../../../src/metrics'
 import { createActiveEntitiesComponent } from '../../../src/ports/activeEntities'
@@ -19,7 +19,6 @@ import { createFailedDeployments } from '../../../src/ports/failedDeployments'
 import { createTestDatabaseComponent } from '../../../src/ports/postgres'
 import { createSequentialTaskExecutor } from '../../../src/ports/sequecuentialTaskExecutor'
 import { ContentAuthenticator } from '../../../src/service/auth/Authenticator'
-import { ServiceImpl } from '../../../src/service/ServiceImpl'
 import { EntityVersion } from '../../../src/types'
 import { NoOpServerValidator, NoOpValidator } from '../../helpers/service/validations/NoOpValidator'
 import { NoOpPointerManager } from '../service/pointers/NoOpPointerManager'
@@ -57,78 +56,78 @@ describe('activeEntities', () => {
     })
 
     it(`return a cached result on second retrieval when the same pointer is asked twice using the same case`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, pointers: pointersToUseByCases.lowercase }]))
 
-      const firstResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), undefined, pointersToUseByCases.lowercase)
 
       sut.mockClear()
 
-      const secondResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const secondResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
 
       expect(firstResult).toMatchObject(secondResult)
     })
 
     it(`return a cached result on second retrieval when the same pointer is asked twice using the same case but the pointers retrieved are in different case`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, pointers: pointersToUseByCases.uppercase }]))
 
-      const firstResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), undefined, pointersToUseByCases.lowercase)
 
       sut.mockClear()
 
-      const secondResult = await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      const secondResult = await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).not.toHaveBeenCalled()
 
       expect(firstResult).toMatchObject(secondResult)
     })
 
     it(`return a cached result on second retrieval when the same pointer is asked twice using different cases (1st lower, then upper)`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, pointers: pointersToUseByCases.lowercase }]))
 
-      const firstResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), undefined, pointersToUseByCases.lowercase)
 
       sut.mockClear()
 
-      const secondResult = await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      const secondResult = await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).not.toHaveBeenCalled()
 
       expect(firstResult).toMatchObject(secondResult)
     })
 
     it(`return a cached result on second retrieval when the same pointer is asked twice using different cases (1st upper, then lower)`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, pointers: pointersToUseByCases.lowercase }]))
 
-      const firstResult = await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      const firstResult = await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), undefined, pointersToUseByCases.uppercase)
 
       sut.mockClear()
 
-      const secondResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const secondResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
 
       expect(firstResult).toMatchObject(secondResult)
     })
 
     it(`return a cached result after the first retrieval when the same pointer is asked three times using different cases`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, pointers: pointersToUseByCases.lowercase }]))
 
-      const firstResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), undefined, pointersToUseByCases.lowercase)
 
       sut.mockClear()
 
-      const secondResult = await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      const secondResult = await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).not.toHaveBeenCalled()
 
-      const thirdResult = await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      const thirdResult = await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).not.toHaveBeenCalled()
 
       expect(firstResult).toMatchObject(secondResult)
@@ -136,18 +135,18 @@ describe('activeEntities', () => {
     })
 
     it(`return a cached result after the first retrieval when the same pointer is asked three times using different cases inverted`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, pointers: pointersToUseByCases.lowercase }]))
 
-      const firstResult = await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      const firstResult = await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), undefined, pointersToUseByCases.uppercase)
 
       sut.mockClear()
 
-      const secondResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const secondResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
 
-      const thirdResult = await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      const thirdResult = await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
 
       expect(firstResult).toMatchObject(secondResult)
@@ -155,23 +154,23 @@ describe('activeEntities', () => {
     })
 
     it(`return a cached result after the first retrieval when the same pointer is asked four times using different cases`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, pointers: pointersToUseByCases.lowercase }]))
 
       // Call the first time
-      await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       // When a pointer is asked the first time, then the database is reached
       expect(sut).toHaveBeenCalledWith(expect.anything(), undefined, pointersToUseByCases.lowercase)
 
       // Reset spy and call again
       sut.mockClear()
-      await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).not.toHaveBeenCalled()
 
-      await service.components.activeEntities.withPointers(pointersToUseByCases.uppercase)
+      await components.activeEntities.withPointers(pointersToUseByCases.uppercase)
       expect(sut).not.toHaveBeenCalled()
 
-      await service.components.activeEntities.withPointers(pointersToUseByCases.lowercase)
+      await components.activeEntities.withPointers(pointersToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
     })
   })
@@ -187,25 +186,25 @@ describe('activeEntities', () => {
     })
 
     it(`treat entity id on cache keys as case sensitive, then result changes depending on the cases passed in the id`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementation(() => Promise.resolve([{ ...fakeDeployment, entityId: idsToUseByCases.lowercase[0] }]))
 
-      const firstResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), idsToUseByCases.lowercase, undefined)
 
       sut.mockClear()
 
-      const secondResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const secondResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
 
       expect(firstResult).toMatchObject(secondResult)
     })
 
     it(`return a cached result only when the id is equally considering case sensitive (1st lower, 2nd upper, 3rd lower)`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementationOnce(() => Promise.resolve([{ ...fakeDeployment, entityId: idsToUseByCases.lowercase[0] }]))
 
-      const firstResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), idsToUseByCases.lowercase, undefined)
 
       sut.mockClear()
@@ -215,23 +214,23 @@ describe('activeEntities', () => {
         ])
       )
 
-      const secondResult = await service.components.activeEntities.withIds(idsToUseByCases.uppercase)
+      const secondResult = await components.activeEntities.withIds(idsToUseByCases.uppercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), idsToUseByCases.uppercase, undefined)
       expect(secondResult).not.toMatchObject(firstResult)
       expect(secondResult[0].id).toEqual(idsToUseByCases.uppercase[0])
 
       sut.mockClear()
 
-      const thirdResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const thirdResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
       expect(firstResult).toMatchObject(thirdResult)
     })
 
     it(`return a cached result only when the id is equally considering case sensitive (1st lower, 2nd upper, 3rd lower, 4th upper)`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementationOnce(() => Promise.resolve([{ ...fakeDeployment, entityId: idsToUseByCases.lowercase[0] }]))
 
-      const firstResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), idsToUseByCases.lowercase, undefined)
 
       sut.mockClear()
@@ -241,27 +240,27 @@ describe('activeEntities', () => {
         ])
       )
 
-      const secondResult = await service.components.activeEntities.withIds(idsToUseByCases.uppercase)
+      const secondResult = await components.activeEntities.withIds(idsToUseByCases.uppercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), idsToUseByCases.uppercase, undefined)
       expect(secondResult).not.toMatchObject(firstResult)
       expect(secondResult[0].id).toEqual(idsToUseByCases.uppercase[0])
 
       sut.mockClear()
 
-      const thirdResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const thirdResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
       expect(firstResult).toMatchObject(thirdResult)
 
-      const fourthResult = await service.components.activeEntities.withIds(idsToUseByCases.uppercase)
+      const fourthResult = await components.activeEntities.withIds(idsToUseByCases.uppercase)
       expect(sut).not.toHaveBeenCalled()
       expect(secondResult).toMatchObject(fourthResult)
     })
 
     it(`return a cached result only when the id is equally considering case sensitive (1st lower, 2nd upper, 3rd lower, 4th lower)`, async () => {
-      const service = await buildService()
+      const components = await buildComponents()
       sut.mockImplementationOnce(() => Promise.resolve([{ ...fakeDeployment, entityId: idsToUseByCases.lowercase[0] }]))
 
-      const firstResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const firstResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), idsToUseByCases.lowercase, undefined)
 
       sut.mockClear()
@@ -271,25 +270,25 @@ describe('activeEntities', () => {
         ])
       )
 
-      const secondResult = await service.components.activeEntities.withIds(idsToUseByCases.uppercase)
+      const secondResult = await components.activeEntities.withIds(idsToUseByCases.uppercase)
       expect(sut).toHaveBeenCalledWith(expect.anything(), idsToUseByCases.uppercase, undefined)
       expect(secondResult).not.toMatchObject(firstResult)
       expect(secondResult[0].id).toEqual(idsToUseByCases.uppercase[0])
 
       sut.mockClear()
 
-      const thirdResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const thirdResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
       expect(firstResult).toMatchObject(thirdResult)
 
-      const fourthResult = await service.components.activeEntities.withIds(idsToUseByCases.lowercase)
+      const fourthResult = await components.activeEntities.withIds(idsToUseByCases.lowercase)
       expect(sut).not.toHaveBeenCalled()
       expect(firstResult).toMatchObject(fourthResult)
     })
   })
 })
 
-async function buildService() {
+async function buildComponents() {
   const database = createTestDatabaseComponent()
   database.queryWithValues = () => Promise.resolve({ rows: [], rowCount: 0 })
   database.transaction = () => Promise.resolve()
@@ -322,7 +321,7 @@ async function buildService() {
   const sequentialExecutor = createSequentialTaskExecutor({ logs, metrics })
   const activeEntities = createActiveEntitiesComponent({ database, logs, env, metrics, denylist, sequentialExecutor })
 
-  return new ServiceImpl({
+  return {
     env,
     pointerManager,
     failedDeployments,
@@ -338,5 +337,5 @@ async function buildService() {
     deployedEntitiesBloomFilter,
     activeEntities,
     denylist
-  })
+  }
 }
